requirements.txt
# OncoAgent - Pinned Dependencies for Reproducibility
# Hardware: AMD Instinct MI300X | ROCm 7.2.x
# ================================================

# === Core ML Framework ===
torch>=2.3.0                # ROCm maps device="cuda" to HIP internally
transformers>=4.42.0        # Llama 3.1 support
accelerate>=0.31.0          # Distributed training utilities

# === Fine-Tuning (QLoRA) ===
peft>=0.11.0                # LoRA/QLoRA adapters
bitsandbytes>=0.43.0        # 4-bit quantization (ROCm-compatible build required)
datasets>=2.20.0            # HuggingFace datasets loading
trl>=0.9.0                  # SFT Trainer

# === RAG Engine ===
PyMuPDF>=1.24.0             # PDF extraction (fitz)
pymupdf4llm>=0.0.17         # SOTA: PDF to Markdown with table support
chromadb>=0.5.0             # Local vector database
sentence-transformers>=3.0.0 # BioBERT embeddings + CrossEncoder re-ranking
networkx>=3.0               # SOTA: Knowledge Graph / GraphRAG
requests>=2.31.0            # SOTA: API connectivity (CIViC/ClinicalTrials)

# === Agent Orchestration ===
langgraph>=0.1.0            # Multi-agent state graph — Rule #17
langchain-core>=0.2.0       # Base abstractions for agents

# === Inference Engine ===
vllm>=0.5.0                 # PagedAttention serving — Rule #6

# === UI ===
gradio>=4.36.0              # Frontend interface — Rule #29

# === Utilities ===
python-dotenv>=1.0.0        # .env secret management — Rule #24
huggingface-hub>=0.23.0     # Model/dataset downloads

# === Testing & Dev ===
pytest>=8.0.0               # Test framework
reportlab>=4.0.0            # Synthetic PDF generation for testing
